+++ src/main.rs
@@ -2,7 +2,6 @@
 
 #![feature(proc_macro_hygiene, decl_macro)]
 
-use std::path::PathBuf;
 use std::sync::atomic::AtomicU64;
 use std::sync::Arc;
 
@@ -10,60 +9,34 @@ use log::LevelFilter;
 use simplelog::{ConfigBuilder, TermLogger, TerminalMode, ThreadLogMode};
 use structopt::StructOpt;
 
+use crate::opts::Opts;
 use crate::prelude::*;
 
 mod core;
 mod format;
+mod opts;
 mod prelude;
 mod services;
 mod settings;
 mod web;
 
-#[derive(StructOpt, Debug)]
-#[structopt(name = "my-iot", author, about)]
-struct Opt {
-    /// Show only warnings and errors
-    #[structopt(short = "s", long = "silent", conflicts_with = "verbose")]
-    silent: bool,
-
-    /// Show all log messages
-    #[structopt(short = "v", long = "verbose", conflicts_with = "silent")]
-    verbose: bool,
-
-    /// Database URL
-    #[structopt(long, env = "MYIOT_DB", default_value = "my-iot.sqlite3")]
-    db: String,
-
-    /// Run only the specified service IDs.
-    #[structopt(short = "i", long = "service-id")]
-    service_ids: Option<Vec<String>>,
-
-    /// Setting files
-    #[structopt(parse(from_os_str), env = "MYIOT_SETTINGS", default_value = "my-iot.toml")]
-    settings: Vec<PathBuf>,
-
-    /// Prints version information
-    #[structopt(short = "V", long = "version")]
-    pub version: bool,
-}
-
 /// Entry point.
 fn main() -> Result {
-    let opt: Opt = Opt::from_args();
-    if opt.version {
+    let opts = opts::Opts::from_args();
+    if opts.version {
         // I want to print only the version, without the application name.
         println!("{}", crate_version!());
         return Ok(());
     }
 
-    init_logging(opt.silent, opt.verbose)?;
+    init_logging(&opts)?;
 
     info!("Reading the settings…");
-    let settings = settings::read(opt.settings)?;
+    let settings = settings::read(opts.settings)?;
     debug!("Settings: {:?}", &settings);
 
     info!("Opening the database…");
-    let db = Connection::open_and_initialize(&opt.db)?;
+    let db = Connection::open_and_initialize(&opts.db)?;
 
     info!("Starting services…");
     let message_counter = Arc::new(AtomicU64::new(0));
@@ -72,35 +45,39 @@ fn main() -> Result {
         .send(Message::new("my-iot::start").type_(MessageType::ReadNonLogged))?;
     core::db::thread::spawn(db.clone(), &mut bus)?;
     services::db::Db.spawn("system::db".into(), &mut bus, db.clone())?;
-    services::spawn_all(&settings, &opt.service_ids, &mut bus, &db)?;
+    services::spawn_all(&settings, &opts.service_ids, &mut bus, &db)?;
     bus.spawn()?;
 
     info!("Starting web server on port {}…", settings.http.port);
     web::start_server(&settings, db, message_counter)
 }
 
-fn init_logging(silent: bool, verbose: bool) -> Result {
+fn init_logging(opts: &Opts) -> Result {
+    let mut config_builder = ConfigBuilder::new();
+    config_builder
+        .set_thread_level(LevelFilter::Error)
+        .set_target_level(LevelFilter::Error)
+        .set_location_level(LevelFilter::Debug)
+        .set_thread_mode(ThreadLogMode::Names)
+        .set_time_format_str("%F %T%.3f")
+        .set_time_to_local(true)
+        .add_filter_ignore_str("h2")
+        .add_filter_ignore_str("hyper")
+        .add_filter_ignore_str("launch_")
+        .add_filter_ignore_str("reqwest")
+        .add_filter_ignore_str("rustls");
+    if opts.suppress_log_timestamps {
+        config_builder.set_time_level(LevelFilter::Off);
+    }
     TermLogger::init(
-        if silent {
+        if opts.silent {
             LevelFilter::Warn
-        } else if verbose {
+        } else if opts.verbose {
             LevelFilter::Debug
         } else {
             LevelFilter::Info
         },
-        ConfigBuilder::new()
-            .set_thread_level(LevelFilter::Error)
-            .set_target_level(LevelFilter::Error)
-            .set_location_level(LevelFilter::Debug)
-            .set_thread_mode(ThreadLogMode::Names)
-            .set_time_format_str("%F %T%.3f")
-            .set_time_to_local(true)
-            .add_filter_ignore_str("launch_")
-            .add_filter_ignore_str("rustls")
-            .add_filter_ignore_str("reqwest")
-            .add_filter_ignore_str("h2")
-            .add_filter_ignore_str("hyper")
-            .build(),
+        config_builder.build(),
         TerminalMode::Stderr,
     )?;
     Ok(())